gen2-cloud-function/requirements.txt
flask==3.1.0
flask-sqlalchemy~=2.5.1
MarkupSafe==2.1.3
google-auth==2.32.0
functions-framework==3.8.0
cloud-sql-python-connector[pg8000]==1.17.0
sqlalchemy==2.0.38